style: fix ruff 0.16 lint findings before Phase B.2 scan
Mechanical, behavior-preserving changes only: use itertools.pairwise()
instead of zip(x, x[1:]), import typing names from collections.abc, iterate
dicts with .items(), drop an unused import and a redundant int() cast, and
sort one test import block. No scientific parameters, thresholds, frozen
MTFA settings, or lock boundaries were changed.

+++ src/whole_eye_mvp/domain.py
@@ -217,7 +217,7 @@ class B0LockMtfaSettings:
         ratio = self.q_lock_max_cycles_per_mm / step
         if not math.isclose(ratio, round(ratio), rel_tol=0.0, abs_tol=1.0e-12):
             raise ValueError("B0 MTFA frequency step must divide Q-lock maximum exactly")
-        return tuple(round(index * step, 10) for index in range(int(round(ratio)) + 1))
+        return tuple(round(index * step, 10) for index in range(round(ratio) + 1))
 
 
 # Legacy v1 provenance: the first Phase B design used Huygens MTF and stopped on
